Patch 0121
・AthenaProject2278からpacket_db.txtを更新、2007-02-12aSakexeに対応
・skill_unit_move_unit_group() を少し修正（skill.c）
・pid_create() の警告抑制（core.c）

+++ src/common/core.c
@@ -76,7 +76,7 @@ void pid_create(const char* file) {
 	fp = fopen(pid_file,"w");
 	if(fp) {
 #ifdef _WIN32
-		fprintf(fp,"%d",GetCurrentProcessId());
+		fprintf(fp,"%u",(unsigned long)GetCurrentProcessId());
 #else
 		fprintf(fp,"%d",getpid());
 #endif
